Added image integration from database
