+++ public/js/main.js
@@ -3,6 +3,7 @@ import { getUserPosts } from './posts/create-user-posts';
 import { createViewPost } from './posts/create-view-post-panel';
 import { createEditPostPanel } from './posts/create-edit-post-panel';
 import { editPost, validateEditPost } from './posts/edit-post';
+import { uploadPostPhoto, enterPhotos } from './upload-images';
 
 /* global $ axios*/
 
@@ -10,41 +11,21 @@ $(function() {
 
   const state = {};
 
-  function uploadPostPhoto(id) {
-    console.log(state);
-    $.ajax({
-      url: `/posts/upload/${id}`,
-      type: 'POST',
-      data: state.formData,
-      processData: false,
-      contentType: false,
-      success: function(data){
-          console.log('upload successful!\n' + data);
-      },
-      error: err => {
-        console.log(`--- ${JSON.stringify(err)}`);
-      },
-      xhr: function() {
-        // create an XMLHttpRequest
-        var xhr = new XMLHttpRequest();
-        // listen to the 'progress' event
-        xhr.upload.addEventListener('progress', function(evt) {
-          if (evt.lengthComputable) {
-            // calculate the percentage of upload completed
-            var percentComplete = evt.loaded / evt.total;
-            percentComplete = parseInt(percentComplete * 100);
-            // update the Bootstrap progress bar with the new percentage
-            $('.progress-bar').text(percentComplete + '%');
-            $('.progress-bar').width(percentComplete + '%');
-            // once the upload reaches 100%, set the progress bar text to done
-            if (percentComplete === 100) {
-              $('.progress-bar').html('Done');
-            }
-          }
-        }, false);
+  let openFromSide = (arg) => {
+    if($(window).width() < 601){
+      $(arg).animate({
+        width:"100vw"
+      });
+    } else {
+      $(arg).animate({
+        width:"50vw"
+      });
+    }
+  }
 
-        return xhr;
-      }
+  let closeSidePullOut = (arg) => {
+    $(arg).animate({
+      width:"0"
     });
   }
 
@@ -56,22 +37,8 @@ $(function() {
   });
 
   $('#edit-post').on('change', '#upload-photo', function(){
-  var files = $(this).get(0).files;
-  console.log(`Files: ${files}`);
-  if (files.length > 0){
-    // create a FormData object which will be sent as the data payload in the
-    // AJAX request
-    var formData = new FormData();
-    // loop through all the selected files and add them to the formData object
-    for (var i = 0; i < files.length; i++) {
-      var file = files[i];
-      console.log(`File loop: ${file}`);
-      // add the files to formData object for the data payload
-      formData.append('uploads[]', file, file.name);
-    }
-    state.formData = formData;
-    console.log(state.formData.uploads);
-  }
+    const that = $(this);
+    enterPhotos(that);
 });
 
   $('#edit-post').on('click', '#js-edit-post', function(e) {
@@ -87,24 +54,6 @@ $(function() {
     }
     closeSidePullOut('#edit-post');
   });
-
-  function openFromSide(arg) {
-    if($(window).width() < 601){
-      $(arg).animate({
-        width:"100vw"
-      });
-    } else {
-      $(arg).animate({
-        width:"50vw"
-      });
-    }
-  }
-
-  function closeSidePullOut(arg) {
-    $(arg).animate({
-      width:"0"
-    });
-  }
 
   //Event Handlers
   $(".dropdown-button").dropdown();
@@ -137,7 +86,7 @@ $(function() {
     let reference = $(this).attr('href');
     closeSidePullOut(reference);
   });
-
+  //New Post Button Animation
   $('.new-post-button').hover(function() {
     if($(window).width() > 600){
       $('.new-post-button-container').toggleClass('elongated');
@@ -146,24 +95,16 @@ $(function() {
     }
   });
 
-  if($('img .materialboxed .initialized').hasClass('active')){
-    $('.navbar-fixed').hide();
-  }
-
-// $('.side-pull-out').on('click', 'img', function() {
-//     $('.navbar-fixed').toggleClass('hidden');
-// })
-
-// $('body').click(function() {
-//   let count = 1;
-//   $('body').click(function() {
-//     if($('.navbar-fixed').css('display') == 'none'){
-//     $('.navbar-fixed').show();
-//     }
-//   });
-// });
+  // $('#view-post').on('click', 'img.materialboxed', (e) => {
+  //   e.stopPropagation();
+  //   console.log(e.currentTarget);
+  //   console.log($(this));
+  //   setTimeout( () => {
+  //     if($(e.currentTarget).hasClass('initialized')){
+  //       $('.navbar-fixed').height('0px');
+  //     }
+  //   }, 100);
+  // });
 
   //End
-
-  //window.dispatchEvent(new Event('resize'));
 });
