+++ views/index.hbs
@@ -140,20 +140,20 @@
     <div class="row">
       <div class="col s12" id="parent-div">
         <ul class="tabs transparent">
-          <li class="tab col s3"><a id="my-posts-tab" href="#my-posts">My Posts</a></li>
+          <li class="tab col s3"><a id="user-posts-tab" href="#user-posts">My Posts</a></li>
           <li class="tab col s3" id="js-get-all-posts"><a href="#all-posts">All Posts</a></li>
         </ul>
       </div>
     </div>
     <div class="row">
-      <div id="my-posts">
+      <div id="user-posts">
         <div class="new-post-button-container">
           <div class="new-post-button waves-effect waves-light red">
             <a class="white-text"><span class="new-post-button-font">Create New</span><i class="material-icons new-post-button-icon">add</i></a>
           </div>
         </div>
 
-        <div class="post" id="js-my-post"></div>
+        <div class="post" id="js-user-posts"></div>
       </div>
       <div class="side-pull-out full-height absolute-right white z-depth-2" id="edit-post"></div>
     </div>
