+++ views/layouts/layout.hbs
@@ -2,25 +2,25 @@
 
 <html>
 
-<head>
-  <!--Import Google Icon Font-->
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <!--Import materialize.css-->
-  <link type="text/css" rel="stylesheet" href="css/materialize.min.css" media="screen,projection" />
-  <link type="text/css" rel="stylesheet" href="css/main.css" media="screen,projection" />
+  <head>
+    <!--Import Google Icon Font-->
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <!--Import materialize.css-->
+    <link type="text/css" rel="stylesheet" href="css/materialize.min.css" media="screen,projection" />
+    <link type="text/css" rel="stylesheet" href="css/main.css" media="screen,projection" />
 
-  <!--Let browser know website is optimized for mobile-->
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-</head>
+    <!--Let browser know website is optimized for mobile-->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  </head>
 
-<body {{#if loginClass }} class="light-blue" {{/if}}>
-  {{{body}}}
-  <!--Import jQuery before materialize.js-->
-  <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-  <script type="text/javascript" src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
-  <script type="text/javascript" src="js/materialize.min.js"></script>
-  <script defer src="/js/dist/bundle.js"></script>
-  <!-- <script type="text/javascript" src="js/main.js"></script> -->
-</body>
+  <body {{#if loginClass }} class="light-blue" {{/if}}>
+    {{{body}}}
+    <!--Import jQuery before materialize.js-->
+    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
+    <script type="text/javascript" src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
+    <script type="text/javascript" src="js/materialize.min.js"></script>
+    <script defer src="/js/dist/bundle.js"></script>
+    <!-- <script type="text/javascript" src="js/main.js"></script> -->
+  </body>
 
 </html>
